#8 feat : 내 정보 가져오기 + 작성한 글 리스트 가져오기 완료. 로그 작성. 커스텀 예외 처리 삭제.

+++ backend/starcast/src/main/java/com/mobyeoldol/starcast/member/domain/Auth.java
@@ -22,7 +22,7 @@ public class Auth {
     private String authId;
 
     @Column(name = "kakao_uid", nullable = false)
-    private long kakaoUid;
+    private int kakaoUid;
 
     @CreationTimestamp
     @Column(name = "created_date")
